split gmx helpders code
